# Add specialized copy method