--- a/perf/runbench.jl
+++ b/perf/runbench.jl
@@ -3,6 +3,7 @@ using LinearAlgebra
 using Random
 using GLPK
 using MathOptInterface
+# using ProfileView
 const MOI = MathOptInterface
 
 using TimerOutputs
@@ -110,7 +111,7 @@ function time_build_and_solve(to_build, to_solve, At, b, c, scalar = true)
     if to_build !== to_solve
         @timeit "copy" MOI.copy_to(to_solve, to_build, copy_names = false)
     end
-    MOI.set(to_solve, MOI.TimeLimitSec(), 0.10)
+    MOI.set(to_solve, MOI.TimeLimitSec(), 0.0010)
     @time @timeit "opt" MOI.optimize!(to_solve)
     val = MOI.get(to_solve, MOI.SolveTime())
     println(val)
@@ -122,22 +123,26 @@ function solve_GLPK(seed, data; time_limit_sec=Inf)
 
     reset_timer!()
 
+    @timeit "data"  At, b, c = random_data(1, data)
     for i in 1:seed
-
-        @timeit "data"  At, b, c = random_data(i, data)
-
+        # mod(i,5) == 0 && GC.gc()
+        GC.gc()
         bridged_cache, pure_solver = bridged_cache_and_solver()
         @timeit "bc + s" time_build_and_solve(bridged_cache, pure_solver, At, b, c)
-
+        
+        GC.gc()
         cache, pure_solver2 = cache_and_solver()
         @timeit "c + s" time_build_and_solve(cache, pure_solver2, At, b, c)
-
+        
+        GC.gc()
         full_solver = bridged_cached_solver()
         @timeit "bcs" time_build_and_solve(full_solver, full_solver, At, b, c)
-
+        
+        GC.gc()
         full_solver = bridged_cached_solver()
         @timeit "bcs + v" time_build_and_solve(full_solver, full_solver, At, b, c, false)
-
+        
+        GC.gc()
         cache_solver = cached_solver()
         @timeit "cs" time_build_and_solve(cache_solver, cache_solver, At, b, c)
 
@@ -147,4 +152,5 @@ function solve_GLPK(seed, data; time_limit_sec=Inf)
 
 end
 
-solve_GLPK(100, RandomLP(10000, 10000, 0.005); time_limit_sec=5)
+solve_GLPK(2, RandomLP(11, 11, 0.5); time_limit_sec=5)
+solve_GLPK(20, RandomLP(10000, 10000, 0.005); time_limit_sec=5)