Following the lines of https://github.com/jump-dev/Clp.jl/pull/94

Add a batch copy, it seems that the batch is not much better than one-by-one for GLPK.

Running the runbench.jl file from the perf folder.

Before this PR we had:
```
 ──────────────────────────────────────────────────────────────────
                           Time                   Allocations
                   ──────────────────────   ───────────────────────
 Tot / % measured:       143s / 100%            30.4GiB / 100%

 Section   ncalls     time   %tot     avg     alloc   %tot      avg
 ──────────────────────────────────────────────────────────────────
 bcs + v      100    32.9s  23.0%   329ms   7.23GiB  23.8%  74.0MiB
   build      100    18.8s  13.1%   188ms   7.22GiB  23.7%  73.9MiB
   opt        100    13.8s  9.66%   138ms   7.64MiB  0.02%  78.2KiB
 cs           100    29.0s  20.3%   290ms   5.28GiB  17.4%  54.1MiB
   build      100    15.0s  10.5%   150ms   5.28GiB  17.3%  54.0MiB
   opt        100    13.7s  9.61%   137ms   7.64MiB  0.02%  78.2KiB
 bc + s       100    27.8s  19.4%   278ms   5.83GiB  19.2%  59.7MiB
   opt        100    13.6s  9.53%   136ms     0.00B  0.00%    0.00B
   copy       100    10.8s  7.52%   108ms   3.81GiB  12.5%  39.1MiB
   build      100    3.12s  2.19%  31.2ms   2.01GiB  6.62%  20.6MiB
 bcs          100    26.5s  18.6%   265ms   5.28GiB  17.4%  54.1MiB
   opt        100    13.8s  9.63%   138ms   7.64MiB  0.02%  78.2KiB
   build      100    12.5s  8.72%   125ms   5.28GiB  17.3%  54.0MiB
 c + s        100    24.8s  17.3%   248ms   5.27GiB  17.3%  53.9MiB
   opt        100    13.6s  9.50%   136ms     0.00B  0.00%    0.00B
   copy       100    9.07s  6.34%  90.7ms   3.25GiB  10.7%  33.3MiB
   build      100    1.88s  1.32%  18.8ms   2.01GiB  6.62%  20.6MiB
 data         100    2.00s  1.40%  20.0ms   1.54GiB  5.06%  15.8MiB
 ──────────────────────────────────────────────────────────────────
```
After:
```
 ──────────────────────────────────────────────────────────────────
                           Time                   Allocations
                   ──────────────────────   ───────────────────────
 Tot / % measured:       138s / 100%            28.7GiB / 100%

 Section   ncalls     time   %tot     avg     alloc   %tot      avg
 ──────────────────────────────────────────────────────────────────
 bcs + v      100    31.0s  22.4%   310ms   7.23GiB  25.2%  74.0MiB
   build      100    16.9s  12.3%   169ms   7.22GiB  25.2%  73.9MiB
   opt        100    13.7s  9.94%   137ms   7.64MiB  0.03%  78.2KiB
 cs           100    30.6s  22.2%   306ms   5.28GiB  18.4%  54.1MiB
   build      100    16.6s  12.0%   166ms   5.28GiB  18.4%  54.0MiB
   opt        100    13.7s  9.91%   137ms   7.64MiB  0.03%  78.2KiB
 bcs          100    26.4s  19.1%   264ms   5.28GiB  18.4%  54.1MiB
   opt        100    13.7s  9.93%   137ms   7.64MiB  0.03%  78.2KiB
   build      100    12.3s  8.94%   123ms   5.28GiB  18.4%  54.0MiB
 bc + s       100    24.8s  18.0%   248ms   4.82GiB  16.8%  49.3MiB
   opt        100    13.1s  9.51%   131ms     0.00B  0.00%    0.00B
   copy       100    9.08s  6.58%  90.8ms   2.80GiB  9.78%  28.7MiB
   build      100    2.34s  1.69%  23.4ms   2.01GiB  7.03%  20.6MiB
 c + s        100    22.7s  16.5%   227ms   4.50GiB  15.7%  46.1MiB
   opt        100    13.1s  9.51%   131ms     0.00B  0.00%    0.00B
   copy       100    7.29s  5.28%  72.9ms   2.49GiB  8.69%  25.5MiB
   build      100    2.07s  1.50%  20.7ms   2.01GiB  7.03%  20.6MiB
 data         100    2.58s  1.87%  25.8ms   1.54GiB  5.37%  15.8MiB
 ──────────────────────────────────────────────────────────────────
```

I haven't used the profiler though, there might be some gains.
Moreover, https://github.com/jump-dev/MathOptInterface.jl/pull/1122 could help here.