@joaquimg 2020-07-11
If we skip MOIU.canonical, we get:
```
 ──────────────────────────────────────────────────────────────────
                           Time                   Allocations
                   ──────────────────────   ───────────────────────
 Tot / % measured:       137s / 100%            26.9GiB / 100%

 Section   ncalls     time   %tot     avg     alloc   %tot      avg
 ──────────────────────────────────────────────────────────────────
 bcs + v      100    31.5s  23.1%   315ms   7.23GiB  26.9%  74.0MiB
   build      100    17.5s  12.8%   175ms   7.22GiB  26.9%  73.9MiB
   opt        100    13.7s  10.0%   137ms   7.64MiB  0.03%  78.2KiB
 cs           100    28.5s  20.9%   285ms   5.28GiB  19.6%  54.1MiB
   build      100    14.6s  10.7%   146ms   5.28GiB  19.6%  54.0MiB
   opt        100    13.6s  10.0%   136ms   7.64MiB  0.03%  78.2KiB
 bcs          100    26.4s  19.3%   264ms   5.28GiB  19.6%  54.1MiB
   opt        100    13.7s  10.0%   137ms   7.64MiB  0.03%  78.2KiB
   build      100    12.4s  9.07%   124ms   5.28GiB  19.6%  54.0MiB
 bc + s       100    26.2s  19.2%   262ms   3.93GiB  14.6%  40.3MiB
   opt        100    13.1s  9.61%   131ms     0.00B  0.00%    0.00B
   copy       100    8.48s  6.21%  84.8ms   1.92GiB  7.13%  19.6MiB
   build      100    4.30s  3.15%  43.0ms   2.01GiB  7.49%  20.6MiB
 c + s        100    21.9s  16.0%   219ms   3.62GiB  13.5%  37.1MiB
   opt        100    13.1s  9.59%   131ms     0.00B  0.00%    0.00B
   copy       100    6.70s  4.90%  67.0ms   1.61GiB  5.97%  16.4MiB
   build      100    1.80s  1.32%  18.0ms   2.01GiB  7.49%  20.6MiB
 data         100    2.17s  1.59%  21.7ms   1.54GiB  5.73%  15.8MiB
 ──────────────────────────────────────────────────────────────────
```