@joaquimg 2020-07-11
So there are a few caveats,
This as is now not super useful.
However, considering the contiguous indexing might boost a lot this.