@odow 2020-07-11
One difference is that this is allocating way more than the Clp version. Do we know why?