@joaquimg 2020-07-11
Clp does not have the CleverDict to keep track of data later, that's my current bet.