@joaquimg 2020-07-11
I modified the comparisson scripts:
Before
```
 ──────────────────────────────────────────────────────────────────
                           Time                   Allocations
                   ──────────────────────   ───────────────────────
 Tot / % measured:      36.9s / 49.4%           5.79GiB / 100%

 Section   ncalls     time   %tot     avg     alloc   %tot      avg
 ──────────────────────────────────────────────────────────────────
 bcs + v       20    4.96s  27.2%   248ms   1.44GiB  25.0%  74.0MiB
   build       20    3.98s  21.9%   199ms   1.44GiB  24.9%  73.9MiB
   opt         20    916ms  5.03%  45.8ms   1.53MiB  0.03%  78.2KiB
 bcs           20    3.53s  19.4%   176ms   1.06GiB  18.2%  54.1MiB
   build       20    2.54s  14.0%   127ms   1.05GiB  18.2%  54.0MiB
   opt         20    914ms  5.02%  45.7ms   1.53MiB  0.03%  78.2KiB
 cs            20    3.44s  18.9%   172ms   1.06GiB  18.2%  54.1MiB
   build       20    2.46s  13.5%   123ms   1.05GiB  18.2%  54.0MiB
   opt         20    912ms  5.01%  45.6ms   1.53MiB  0.03%  78.2KiB
 bc + s        20    3.19s  17.5%   159ms   1.17GiB  20.1%  59.7MiB
   copy        20    2.10s  11.5%   105ms    781MiB  13.2%  39.0MiB
   opt         20    843ms  4.63%  42.1ms     0.00B  0.00%    0.00B
   build       20    183ms  1.01%  9.17ms    412MiB  6.95%  20.6MiB
 c + s         20    3.08s  16.9%   154ms   1.05GiB  18.2%  53.9MiB
   copy        20    1.92s  10.6%  96.1ms    666MiB  11.2%  33.3MiB
   opt         20    917ms  5.03%  45.8ms     0.00B  0.00%    0.00B
   build       20    185ms  1.02%  9.25ms    412MiB  6.95%  20.6MiB
 data           1   13.4ms  0.07%  13.4ms   15.8MiB  0.27%  15.8MiB
 ──────────────────────────────────────────────────────────────────
```
After
```
 ──────────────────────────────────────────────────────────────────
                           Time                   Allocations
                   ──────────────────────   ───────────────────────
 Tot / % measured:      36.2s / 47.3%           5.45GiB / 100%

 Section   ncalls     time   %tot     avg     alloc   %tot      avg
 ──────────────────────────────────────────────────────────────────
 bcs + v       20    5.08s  29.6%   254ms   1.44GiB  26.5%  74.0MiB
   build       20    4.07s  23.8%   203ms   1.44GiB  26.5%  73.9MiB
   opt         20    933ms  5.45%  46.7ms   1.53MiB  0.03%  78.2KiB
 cs            20    3.59s  21.0%   179ms   1.06GiB  19.4%  54.1MiB
   build       20    2.58s  15.1%   129ms   1.05GiB  19.4%  54.0MiB
   opt         20    946ms  5.52%  47.3ms   1.53MiB  0.03%  78.2KiB
 bcs           20    3.52s  20.6%   176ms   1.06GiB  19.4%  54.1MiB
   build       20    2.51s  14.6%   125ms   1.05GiB  19.4%  54.0MiB
   opt         20    948ms  5.53%  47.4ms   1.53MiB  0.03%  78.2KiB
 bc + s        20    2.71s  15.8%   136ms   0.97GiB  17.8%  49.7MiB
   copy        20    1.64s  9.57%  81.9ms    582MiB  10.4%  29.1MiB
   opt         20    818ms  4.77%  40.9ms     0.00B  0.00%    0.00B
   build       20    194ms  1.13%  9.71ms    412MiB  7.39%  20.6MiB
 c + s         20    2.21s  12.9%   110ms    924MiB  16.6%  46.2MiB
   copy        20    1.15s  6.73%  57.6ms    512MiB  9.18%  25.6MiB
   opt         20    794ms  4.64%  39.7ms     0.00B  0.00%    0.00B
   build       20    201ms  1.17%  10.0ms    412MiB  7.39%  20.6MiB
 data           1   13.6ms  0.08%  13.6ms   15.8MiB  0.28%  15.8MiB
 ──────────────────────────────────────────────────────────────────
```

Timing are better because I forced GC in between tests.
Main comparison here is the `copy` lines.
When solver is integrated with Cache it is not using the new copy_to, probably passing directly...